Re-make the user registration workflow according to he payment API version 4 ID-296 Re-make the user registration workflow according to he payment API version 4

+++ source/api/models/SubscriptionPayment.js
@@ -16,13 +16,13 @@ module.exports = {
             type: 'string'
         },
         status :{
-            type: 'int'
+            type: 'Integer'
         },
         date:{
             type: 'string'
         },
         amount:{
-            type:'double'
+            type:'float'
         },
         operator:{
             type:'string'
